requirements.txt
Django==5.1.3
django-environ==0.11.2
mysqlclient==2.2.5
djangorestframework==3.15.2
djang-filter==24.3
djangorestframework-simplejwt==5.3.1
